linux-gen: packet: shrink _odp_packet_input_flags_t to 32 bits
Now with fewer packet input flags, shrink _odp_packet_input_flags_t
from 8 to 4 bytes. This reduces the size of packet_parset_t by 8 bytes
since structure padding there gets reduced too. Add explicit unused
padding bytes in packet header (for now) to avoid moving other packet
header fields between cache lines.

+++ platform/linux-generic/include/odp/api/plat/packet_inline_types.h
@@ -60,44 +60,44 @@ extern const _odp_packet_inline_offset_t _odp_packet_inline;
 /* Packet input & protocol flags */
 typedef union {
 	/* All input flags */
-	uint64_t all;
+	uint32_t all;
 
 	/* Individual input flags */
 	struct {
-		uint64_t cls_mark: 1; /* Classifier mark value present*/
+		uint32_t cls_mark: 1; /* Classifier mark value present*/
 
-		uint64_t flow_hash:1; /* Flow hash present */
-		uint64_t timestamp:1; /* Timestamp present */
+		uint32_t flow_hash:1; /* Flow hash present */
+		uint32_t timestamp:1; /* Timestamp present */
 
-		uint64_t l2:1;        /* known L2 protocol present */
-		uint64_t l3:1;        /* known L3 protocol present */
-		uint64_t l4:1;        /* known L4 protocol present */
+		uint32_t l2:1;        /* known L2 protocol present */
+		uint32_t l3:1;        /* known L3 protocol present */
+		uint32_t l4:1;        /* known L4 protocol present */
 
-		uint64_t eth:1;       /* Ethernet */
-		uint64_t eth_bcast:1; /* Ethernet broadcast */
-		uint64_t eth_mcast:1; /* Ethernet multicast */
-		uint64_t jumbo:1;     /* Jumbo frame */
-		uint64_t vlan:1;      /* VLAN hdr found */
-		uint64_t vlan_qinq:1; /* Stacked VLAN found, QinQ */
+		uint32_t eth:1;       /* Ethernet */
+		uint32_t eth_bcast:1; /* Ethernet broadcast */
+		uint32_t eth_mcast:1; /* Ethernet multicast */
+		uint32_t jumbo:1;     /* Jumbo frame */
+		uint32_t vlan:1;      /* VLAN hdr found */
+		uint32_t vlan_qinq:1; /* Stacked VLAN found, QinQ */
 
-		uint64_t arp:1;       /* ARP */
+		uint32_t arp:1;       /* ARP */
 
-		uint64_t ipv4:1;      /* IPv4 */
-		uint64_t ipv6:1;      /* IPv6 */
-		uint64_t ip_bcast:1;  /* IP broadcast */
-		uint64_t ip_mcast:1;  /* IP multicast */
-		uint64_t ipfrag:1;    /* IP fragment */
-		uint64_t ipopt:1;     /* IP optional headers */
+		uint32_t ipv4:1;      /* IPv4 */
+		uint32_t ipv6:1;      /* IPv6 */
+		uint32_t ip_bcast:1;  /* IP broadcast */
+		uint32_t ip_mcast:1;  /* IP multicast */
+		uint32_t ipfrag:1;    /* IP fragment */
+		uint32_t ipopt:1;     /* IP optional headers */
 
-		uint64_t ipsec:1;     /* IPSec packet. Required by the
+		uint32_t ipsec:1;     /* IPSec packet. Required by the
 					   odp_packet_has_ipsec_set() func. */
 
-		uint64_t color:2;     /* Packet color for traffic mgmt */
-		uint64_t nodrop:1;    /* Drop eligibility status */
+		uint32_t color:2;     /* Packet color for traffic mgmt */
+		uint32_t nodrop:1;    /* Drop eligibility status */
 
-		uint64_t l3_chksum_done:1; /* L3 checksum validation done */
-		uint64_t l4_chksum_done:1; /* L4 checksum validation done */
-		uint64_t udp_chksum_zero:1; /* UDP header had 0 as chksum */
+		uint32_t l3_chksum_done:1; /* L3 checksum validation done */
+		uint32_t l4_chksum_done:1; /* L4 checksum validation done */
+		uint32_t udp_chksum_zero:1; /* UDP header had 0 as chksum */
 	};
 
 } _odp_packet_input_flags_t;
